pyproject.toml
[project]
name = "govee"
version = "0.1.0"
description = "CLI tool to control Govee H6046 RGBIC TV Light Bars via Bluetooth"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "you@example.com"}
]
keywords = ["govee", "bluetooth", "ble", "led", "rgb", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Topic :: Home Automation",
]

dependencies = [
    "bleak>=0.21.0",
    "click>=8.0.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
govee = "govee.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/govee"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
]
